refactor(downloader): migrate driver to transmission-rpc
